requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
transformers>=4.21.0
torch>=1.9.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
scikit-learn>=1.3.0
umap-learn>=0.5.3
python-dotenv>=1.0.0
tiktoken>=0.4.0